+++ cmd/repoGetInfo.go
@@ -26,8 +26,9 @@ import (
 	"sync"
 
 	"github.com/J-Siu/go-gitapi/v3/api"
+	"github.com/J-Siu/go-helper/v2/ezlog"
 	"github.com/J-Siu/go-mygit/v3/global"
-	"github.com/J-Siu/go-mygit/v3/lib"
+	"github.com/J-Siu/go-mygit/v3/helper"
 	"github.com/spf13/cobra"
 )
 
@@ -39,7 +40,7 @@ var repoGetInfoCmd = &cobra.Command{
 	Long:    "Get info(json). " + global.TXT_REPO_DIR_LONG + global.TXT_FLAGS_USE,
 	Run: func(cmd *cobra.Command, args []string) {
 		var (
-			out = make(chan api.IApi, 10)
+			out = make(chan *string, 10)
 			wg  sync.WaitGroup
 		)
 		if len(args) == 0 {
@@ -52,14 +53,16 @@ var repoGetInfoCmd = &cobra.Command{
 						property = remote.GitApiProperty(&workPath, global.Flag.Debug)
 						ga       = new(api.Info).New(property).Get()
 					)
-					lib.RepoDoRun(ga, global.Flag.NoParallel, &wg, out)
+					helper.GitApiDoWrapper(ga, &global.Flag, &wg, out)
 				}
 			}
 			wg.Wait()
 			close(out)
 		}()
-		for gitApi := range out {
-			lib.RepoOutput(gitApi, global.Flag, false, false)
+		global.Flag.SingleLine = false
+		global.Flag.StatusOnly = false
+		for o := range out {
+			ezlog.Log().Se().M(o).Out()
 		}
 
 	},