+++ cmd/repoList.go
@@ -23,11 +23,12 @@ THE SOFTWARE.
 package cmd
 
 import (
+	"fmt"
 	"sync"
 
 	"github.com/J-Siu/go-gitapi/v3/api"
 	"github.com/J-Siu/go-mygit/v3/global"
-	"github.com/J-Siu/go-mygit/v3/lib"
+	"github.com/J-Siu/go-mygit/v3/helper"
 	"github.com/spf13/cobra"
 )
 
@@ -40,7 +41,7 @@ var repoListCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		var (
 			// out = make(chan *string, 10)
-			out = make(chan api.IApi, 10)
+			out = make(chan *string, 10)
 			wg  sync.WaitGroup
 		)
 		go func() {
@@ -50,14 +51,16 @@ var repoListCmd = &cobra.Command{
 						property = remote.GitApiProperty(nil, global.Flag.Debug)
 						ga       = new(api.InfoList).New(property, remote.Vendor, p).Get()
 					)
-					lib.RepoDoRun(ga, global.Flag.NoParallel, &wg, out)
+					helper.GitApiDoWrapper(ga, &global.Flag, &wg, out)
 				}
 			}
 			wg.Wait()
 			close(out)
 		}()
-		for gitApi := range out {
-			lib.RepoOutput(gitApi, global.Flag, false, false)
+		global.Flag.SingleLine = false
+		global.Flag.StatusOnly = false
+		for o := range out {
+			fmt.Print(*o)
 		}
 	},
 }