+++ cmd/repoSetDescription.go
@@ -23,11 +23,12 @@ THE SOFTWARE.
 package cmd
 
 import (
+	"fmt"
 	"sync"
 
 	"github.com/J-Siu/go-gitapi/v3/api"
 	"github.com/J-Siu/go-mygit/v3/global"
-	"github.com/J-Siu/go-mygit/v3/lib"
+	"github.com/J-Siu/go-mygit/v3/helper"
 	"github.com/spf13/cobra"
 )
 
@@ -38,7 +39,7 @@ var repoSetDescriptionCmd = &cobra.Command{
 	Short:   "Set description",
 	Run: func(cmd *cobra.Command, args []string) {
 		var (
-			out = make(chan api.IApi, 10)
+			out = make(chan *string, 10)
 			wg  sync.WaitGroup
 		)
 		if len(args) > 0 {
@@ -46,16 +47,17 @@ var repoSetDescriptionCmd = &cobra.Command{
 				for _, remote := range global.Conf.MergedRemotes {
 					var (
 						property = remote.GitApiProperty(nil, global.Flag.Debug)
-						ga       = new(api.Description).New(property).Set()
+						ga       = new(api.Description).New(property).Set(args[0])
 					)
-					ga.Info.Description = args[0]
-					lib.RepoDoRun(ga, global.Flag.NoParallel, &wg, out)
+					helper.GitApiDoWrapper(ga, &global.Flag, &wg, out)
 				}
 				wg.Wait()
 				close(out)
 			}()
-			for gitApi := range out {
-				lib.RepoOutput(gitApi, global.Flag, true, true)
+			global.Flag.SingleLine = true
+			global.Flag.StatusOnly = true
+			for o := range out {
+				fmt.Print(*o)
 			}
 		}
 	},