+++ cmd/repoSetTopic.go
@@ -23,11 +23,12 @@ THE SOFTWARE.
 package cmd
 
 import (
+	"fmt"
 	"sync"
 
 	"github.com/J-Siu/go-gitapi/v3/api"
 	"github.com/J-Siu/go-mygit/v3/global"
-	"github.com/J-Siu/go-mygit/v3/lib"
+	"github.com/J-Siu/go-mygit/v3/helper"
 	"github.com/spf13/cobra"
 )
 
@@ -39,7 +40,7 @@ var repoSetTopicCmd = &cobra.Command{
 	Long:    "Set topic. " + global.TXT_FLAGS_USE,
 	Run: func(cmd *cobra.Command, args []string) {
 		var (
-			out = make(chan api.IApi, 10)
+			out = make(chan *string, 10)
 			wg  sync.WaitGroup
 		)
 		go func() {
@@ -50,13 +51,15 @@ var repoSetTopicCmd = &cobra.Command{
 				)
 				ga.Info.Topics = &args
 				ga.Info.Names = &args
-				lib.RepoDoRun(ga, global.Flag.NoParallel, &wg, out)
+				helper.GitApiDoWrapper(ga, &global.Flag, &wg, out)
 			}
 			wg.Wait()
 			close(out)
 		}()
-		for gitApi := range out {
-			lib.RepoOutput(gitApi, global.Flag, true, true)
+		global.Flag.SingleLine = true
+		global.Flag.StatusOnly = true
+		for o := range out {
+			fmt.Print(*o)
 		}
 	},
 }