+++ lib/remote.go
@@ -24,11 +24,8 @@ package lib
 
 import (
 	"path"
-	"sync"
 
 	"github.com/J-Siu/go-gitapi/v3/base"
-	"github.com/J-Siu/go-gitcmd/v3/gitcmd"
-	"github.com/J-Siu/go-helper/v2/cmd"
 	"github.com/J-Siu/go-helper/v2/file"
 )
 
@@ -51,9 +48,10 @@ type Remote struct {
 }
 
 func (t *Remote) GitApiProperty(workPathP *string, debug bool) *base.Property {
-	var fullPath string = *file.FullPath(workPathP)
-	var repo string = path.Base(fullPath)
-
+	var (
+		fullPath string = *file.FullPath(workPathP)
+		repo     string = path.Base(fullPath)
+	)
 	property := base.Property{
 		Name:       t.Name,
 		Token:      t.Token,
@@ -64,81 +62,5 @@ func (t *Remote) GitApiProperty(workPathP *string, debug bool) *base.Property {
 		Repo:       repo,
 		Debug:      debug,
 	}
-
 	return &property
-}
-
-// Add all Remotes into git repository
-func (t *Remote) Add(workPathP *string) *cmd.Cmd {
-	t.Remove(workPathP)
-	var (
-		fullPath string   = *file.FullPath(workPathP)
-		repo     string   = path.Base(fullPath)
-		git      string   = t.Ssh + ":/" + t.User + "/" + repo + ".git"
-		myCmd    *cmd.Cmd = gitcmd.RemoteAdd(fullPath, t.Name, git)
-	)
-	t.Output <- MyCmdLog(myCmd, workPathP, t.NoTitle)
-	return myCmd
-}
-
-// Remove all Remotes in git repository
-func (t *Remote) Remove(workPathP *string) *cmd.Cmd {
-	var (
-		myCmd *cmd.Cmd
-	)
-	if gitcmd.RemoteExist(*workPathP, t.Name) {
-		myCmd = gitcmd.RemoteRemove(*workPathP, t.Name)
-	}
-	if myCmd != nil {
-		t.Output <- MyCmdLog(myCmd, workPathP, t.NoTitle)
-	}
-	return myCmd
-}
-
-func MyCmdLog(myCmd *cmd.Cmd, workPathP *string, noTitle bool) *string {
-	var (
-		str1  string
-		str2  string
-		title string
-	)
-	if !noTitle {
-		if workPathP == nil {
-			title = myCmd.CmdLn
-		} else {
-			title = *workPathP + ": " + myCmd.CmdLn
-		}
-	}
-	if len(title) > 0 {
-		str2 = title + ":"
-	}
-	if str2 != "" {
-		str1 += str2 + "\n"
-	}
-	str2 = myCmd.Stdout.String()
-	if str2 != "" {
-		str1 += str2
-	}
-	str2 = myCmd.Stderr.String()
-	if str2 != "" {
-		str1 += str2
-	}
-	return &str1
-}
-
-func GitRunWrapper(gitCmd *gitcmd.GitCmd, workPath string, wgP *sync.WaitGroup, noParallel, noTitle bool, out chan *string) {
-	if noParallel {
-		gitRun(gitCmd, workPath, nil, noTitle, out)
-	} else {
-		wgP.Add(1)
-		go gitRun(gitCmd, workPath, wgP, noTitle, out)
-	}
-}
-
-func gitRun(gitCmd *gitcmd.GitCmd, workPath string, wgP *sync.WaitGroup, noTitle bool, out chan *string) *cmd.Cmd {
-	if wgP != nil {
-		defer wgP.Done()
-	}
-	var myCmd = gitCmd.Run()
-	out <- MyCmdLog(myCmd, &workPath, noTitle)
-	return myCmd
 }